DistanceMatrix refactor Phase A+B: inventory, widen accessors to size_t
Phase A (no code changes): confirmed 508 call-site occurrences across
28 files (17 in-scope), close to the plan's original 526/23 estimate.
Found a fourth field=-1-on-unsigned site the plan's original inventory
missed (DistanceMatrix_impl.hpp:59), traced all four such sites
individually and classified them - three are already-safe (overwritten
before any accessor call), one (DistanceRow's stream constructor) is
confirmed dead AND actually broken (DistanceRow never overrides
objInitFromStream, so columns stays wrapped at SIZE_MAX forever) but
also confirmed unused anywhere, so left in place rather than deleted
(a design decision outside this refactor's stated scope). Also
confirmed the four out-of-scope caller files aren't wired into any
CMake target at all - not just unverifiable but entirely uncompiled,
so there's no actual compile-break risk from the signature change to
guard against there; left untouched rather than making blind edits to
dead code.

Phase B: widened getDistance/setDistance/getIdentifier/setIdentifier
from int to size_t across DistanceMatrix/FloatDistanceMatrix/
DistanceRow - header-only change. Zero call-site changes needed
anywhere in the active build (0 compile errors on rebuild) - every
in-scope caller already used size_t-typed loop variables, confirmed
directly in NeighborJoining.hpp (136 call sites, all already size_t).
Measured via a clang-tidy narrowing-conversions before/after diff
across the same 57 in-scope files: 285 -> 170 findings, 115 resolved
by this header-only change alone.

Verified: full rebuild (0 errors), ctest (2/2), RunExamples.sh
(byte-identical).

+++ include/fastphylo/core/DistanceRow.hpp
@@ -52,11 +52,11 @@ public:
   
   //---------------------------------
   //GET AND SET DISTANCE
-  DistanceType getDistance(int j) const{
+  DistanceType getDistance(size_t j) const{
       return D[j];
   };
-  
-  void setDistance(int j, DistanceType d) {
+
+  void setDistance(size_t j, DistanceType d) {
       D[j] = d;
   };
